services/marketplace/requirements.txt
fastapi
uvicorn
pydantic-settings
cryptography>=44.0.0
redis
